mathworks: Update catalog to specify FSBL per board
Previously, FSBL was specified per app, and on Zynq, a separate
BOOT.BIN file was generated per app. To begin removing the dependency
from bootgen and enable future use of the SPL as the FSBL, start
generating only a single BOOT.BIN per catalog.

On Zynq, instrad of baking the bitstream into BOOT.BIN, simply copy it
to the SD card, and set the default bitstream as "system.bit"

On Altera, no build changes.

On both platforms, the FSBL/Handoff files must be specified within the
default node, instead of within the app node.