mke2img: increase nb_blocks to avoid genext2fs error
On some host filesystems when using mke2img to generate a rootfs image,
we see the error:

genext2fs: couldn't allocate a block (no free space)

This may be due to mapping between the host filesystem's disk usage and
ext2/3/4. Use the --aparent-size option to find the largest of the
aparent and actual disk usage, and then use that for generating the
filesystem image.

Signed-off-by: Matthew Fornero <mfornero@mathworks.com>