mathworks: common: Enable use of ccache
ccache is a mechansim to cache build artifacts generated by the same
source / toolchain / build flags. It can speed up builds for similar
boards by reusing the artifacts instead of recompiling them.

Enable this via build.py with the --ccache flag.

When changing toolchain versions, it is a good idea to flush
the cache on the first build by enabling the --ccache and
--ccache-clean flags.