gitlab: Enable arbitrary buildroot variables
Enable arbitrary setting of BR2_xxx variables via CI automation.

BR2_xxx variables can be set as follows:
* CONFIG_BR2_xxx=YYY sets a variable for all builds
* CONFIG_[MASK]_BR2_xxx=YYY sets a variable for all builds matching
  [MASK]

Where [MASK] is one of the value of:
* $CONFIG_PLATFORM
* $CONFIG_PROJECT
* ${CONFIG_PLATFORM}_${CONFIG_PROJECT}
* The board name as discovered from the board directory

In all cases, the [MASK] is specified in uppercase, regardless of
the case of the CONFIG variable.

For example:
* CONFIG_ZYNQ_BR2_xxx=YYY would set BR2_xxx=YYY for all builds with
  CONFIG_PLATFORM=zynq
* CONFIG_ZYNQSDR_BR2_xxx=YYY would set BR2_xxx=yyy for all builds with
  CONFIG_PROJECT=zynqsdr
* CONFIG_ZYNQ_ZYNQSDR_BR2_xxx=YYY would set BR2_xxx=yyy for all builds
* with CONFIG_PLATFORM=zynq and CONFIG_PROJECT=zynqsdr

This replaces the Linux / Uboot specific variables previously supported.